Some refactoring and fixed datatest ids
