+++ shell/machine-config/google.vue
@@ -10,19 +10,16 @@ import LabeledSelect from '@shell/components/form/LabeledSelect';
 import { LabeledInput } from '@components/Form/LabeledInput';
 import { Checkbox } from '@components/Form/Checkbox';
 import ArrayList from '@shell/components/form/ArrayList';
-
-import { NORMAN } from '@shell/config/types';
+import GCEImage from '@shell/machine-config/components/GCEImage.vue';
 import { convertStringToKV, convertKVToString } from '@shell/utils/object';
 import KeyValue from '@shell/components/form/KeyValue';
 import {
-  getGKEZones, getGKEImageFamilies, getGKEFamiliesFromProject, getGKEDiskTypes, getGKENetworks, getGKEMachineTypes, getGKESubnetworks, getGKESharedSubnetworks
+  getGKEZones, getGKEDiskTypes, getGKENetworks, getGKEMachineTypes, getGKESubnetworks, getGKESharedSubnetworks
 } from '@pkg/gke/util/gcp';
 import { formatSharedNetworks, formatNetworkOptions, formatSubnetworkOptions } from '@pkg/gke/util/formatter';
 import { mapGetters } from 'vuex';
 import { sortBy, sortableNumericSuffix } from '@shell/utils/sort';
-import debounce from 'lodash/debounce';
 const GKE_NONE_OPTION = 'none';
-const DEFAULT_PROJECTS = 'suse-cloud, ubuntu-os-cloud';
 
 const DEFAULT_MIN_DISK = 10;
 
@@ -54,6 +51,7 @@ export default {
     LabeledInput,
     LabeledSelect,
     Loading,
+    GCEImage
   },
 
   mixins: [CreateEditView, FormValidation],
@@ -75,10 +73,7 @@ export default {
       type:     String,
       required: true,
     },
-    poolId: {
-      type:     String,
-      required: true,
-    },
+
     disabled: {
       type:    Boolean,
       default: false
@@ -88,14 +83,6 @@ export default {
   async fetch() {
     if ( !this.credentialId ) {
       return;
-    }
-
-    try {
-      if ( this.credential?.id !== this.credentialId ) {
-        this.credential = await this.$store.dispatch('rancher/find', { type: NORMAN.CLOUD_CREDENTIAL, id: this.credentialId });
-      }
-    } catch (e) {
-      this.credential = null;
     }
 
     for (const key in this.defaultConfig) {
@@ -110,39 +97,20 @@ export default {
   data() {
     return {
       defaultConfig,
-      credential:          null,
-      images:              {},
-      // loading:               false,
       loadingZones:        false,
       loadingDiskTypes:    false,
       loadingNetworks:     false,
-      loadingImages:       false,
       loadingMachineTypes: false,
-      loadingFamilies:     false,
       zones:               [],
-      families:            {},
-      machineImages:       [],
       diskTypes:           [],
       networks:            [],
       subnetworks:         [],
       sharedSubnetworks:   [],
       machineTypes:        [],
-      imageProjects:       DEFAULT_PROJECTS,
-      showDeprecated:      false,
-      family:              null,
       useIpAliases:        false,
       minDisk:             DEFAULT_MIN_DISK,
-      defaultProjects:     DEFAULT_PROJECTS,
+      imageAreaValid:      true,
       fvFormRuleSets:      [
-        {
-          path: 'imageProjects', rootObject: this, rules: ['required', 'projects']
-        },
-        {
-          path: 'family', rootObject: this, rules: ['required']
-        },
-        {
-          path: 'machineImage', rootObject: this, rules: ['required']
-        },
         { path: 'diskType', rules: ['required'] },
         { path: 'diskSize', rules: ['required', 'isPositive', 'minDiskSize'] },
         { path: 'machineType', rules: ['required'] },
@@ -151,7 +119,6 @@ export default {
     };
   },
   created() {
-    this.debouncedLoadFamilies = debounce(this.getFamilies, 500);
     if (this.mode === _CREATE) {
       this.$emit('validationChanged', false);
       this.value.project = this.projectId;
@@ -164,7 +131,6 @@ export default {
     } else {
       this.value.setInternalFirewallRulePrefix = !!this.value.internalFirewallRulePrefix;
       this.value.setExternalFirewallRulePrefix = !!this.value.externalFirewallRulePrefix;
-      this.imageProjects = `${ this.getProjectFromImage() }`;
     }
   },
 
@@ -172,8 +138,8 @@ export default {
     credentialId() {
       this.$fetch();
     },
-    fvFormIsValid(newValue) {
-      this.$emit('validationChanged', !!newValue);
+    validationProperties(newVal) {
+      this.$emit('validationChanged', !!newVal);
     },
 
     'value.zone'() {
@@ -191,17 +157,7 @@ export default {
         this.value.openPort.push('6443');
       }
     },
-    family(neu) {
-      this.getImages(neu);
-    },
 
-    'imageProjects'() {
-      this.debouncedLoadFamilies();
-    },
-
-    showDeprecated() {
-      this.getFamilies();
-    },
     networkOptions(neu) {
       if (neu && neu.length && !this.value.network) {
         const defaultNetwork = neu.find((network) => network?.name === 'default');
@@ -243,73 +199,10 @@ export default {
     project() {
       return this.value.project;
     },
-    familyOptions() {
-      const out = [];
-
-      Object.keys(this.families).forEach((groupLabel) => {
-        const instances = this.families[groupLabel];
-
-        const groupOption = { label: groupLabel, kind: 'group' };
-        const familyOptions = instances.map((instance) => {
-          return {
-            value: { family: instance, project: groupLabel },
-            label: instance,
-            group: groupLabel
-          };
-        });
-
-        out.push(groupOption);
-        out.push(...familyOptions);
-      });
-
-      return out;
+    validationProperties() {
+      return this.fvFormIsValid && this.imageAreaValid;
     },
-    imageOptions() {
-      let out = [];
 
-      if (!this.showDeprecated) {
-        out = this.machineImages.map((image) => {
-          const value = {
-            name:     image.name,
-            selfLink: image.selfLink,
-            diskSize: image.diskSizeGb
-          };
-          const deprecated = !!image?.deprecated;
-
-          return {
-            value,
-            label: !deprecated ? this.t('cluster.machineConfig.gce.machineImage.option', { name: image.name, description: image.description }) : this.t('cluster.machineConfig.gce.machineImage.deprecatedOption', { name: image.name, description: image.description }),
-          };
-        });
-      } else {
-        const deprecatedImages = [];
-        const activeImages = [];
-
-        this.machineImages.forEach((image) => {
-          const value = {
-            name:     image.name,
-            selfLink: image.selfLink,
-            diskSize: image.diskSizeGb
-          };
-          const deprecated = !!image?.deprecated;
-
-          if (!deprecated) {
-            activeImages.push({
-              value,
-              label: !deprecated ? this.t('cluster.machineConfig.gce.machineImage.option', { name: image.name, description: image.description }) : this.t('cluster.machineConfig.gce.machineImage.deprecatedOption', { name: image.name, description: image.description }),
-            });
-          } else {
-            deprecatedImages.push({
-              value,
-              label: !deprecated ? this.t('cluster.machineConfig.gce.machineImage.option', { name: image.name, description: image.description }) : this.t('cluster.machineConfig.gce.machineImage.deprecatedOption', { name: image.name, description: image.description }),
-            });
-          }
-          out = [...activeImages, ...deprecatedImages];
-        });
-      }
-
-      return out;
-    },
     sharedNetworks() {
       return formatSharedNetworks(this.sharedSubnetworks);
     },
@@ -322,15 +215,7 @@ export default {
     subnetworkOptions() {
       return formatSubnetworkOptions(this.t, this.value.network, this.subnetworks, this.sharedNetworks, this.useIpAliases);
     },
-    machineImage: {
-      get() {
-        return this.value?.machineImage ? this.getImageNameFromImage() : '';
-      },
-      set(neu) {
-        this.setMachineImage(neu);
-        this.diskSize = neu?.diskSize ? neu.diskSize : 10;
-      }
-    },
+
     selectedNetwork: {
       get() {
         const { network } = this.value;
@@ -403,7 +288,6 @@ export default {
       try {
         const res = await getGKEZones(this.$store, this.credentialId, this.project, {});
 
-        // TODO Should consider moving this to the util function and check if we need regions similar to GKE
         this.zones = sortBy((res.items || []).map((z) => {
           z.disabled = z?.status?.toLowerCase() !== 'up';
           z.sortName = sortableNumericSuffix(z.name);
@@ -415,110 +299,6 @@ export default {
 
         return '';
       }
-    },
-    getProjectFromImage() {
-      if (!!this.value?.machineImage) {
-        return this.value.machineImage.split('/')[0];
-      }
-    },
-    getImageNameFromImage() {
-      if (!!this.value?.machineImage) {
-        return this.value.machineImage.split('/')[3];
-      }
-    },
-
-    async getFamilies() {
-      this.loadingFamilies = true;
-      let out = {};
-
-      if (this.imageProjects) {
-        try {
-          let projects = this.imageProjects.replace(/ /g, '');
-
-          if (projects.endsWith(',')) {
-            projects = projects.substring(0, projects.length - 1 );
-          }
-
-          if (!!projects) {
-            out = await getGKEFamiliesFromProject(this.$store, this.credentialId, this.value.project, this.location, projects, false) || {};
-
-            if (!!this.showDeprecated) {
-              const deprecatedResponse = await getGKEFamiliesFromProject(this.$store, this.credentialId, this.project, this.location, projects, true);
-
-              Object.keys(deprecatedResponse).forEach((family) => {
-                if ( !out[family]) {
-                  out[family] = family;
-                } else {
-                  out[family] = [...new Set([...out[family], ...deprecatedResponse[family]])];
-                }
-              });
-            }
-          }
-        } catch (e) {
-          // This fails often if user mistypes, so it's better to swallow the error instead of spamming with errors
-          this.errors.push(e.data);
-
-          return '';
-        }
-      }
-      this.families = out;
-
-      // When editing existing cluster, we need to find corresponding family
-      if (!!this.value.machineImage) {
-        let found = false;
-
-        const project = this.getProjectFromImage();
-        const families = out[project] || [];
-        const imageName = this.getImageNameFromImage();
-
-        let count = 0;
-
-        while (!found && count < families.length) {
-          const images = await this.getImagesInProject({ family: families[count], project }, true );
-          const filtered = images.filter((image) => image.name === imageName);
-
-          if (filtered.length > 0) {
-            found = true;
-            this.family = { family: filtered[0].family, project };
-          }
-          count += 1;
-        }
-      }
-      // If we had to reload list of families, we need to reset selected family if it is no longer in the list
-      if ( !!this.family?.project && !this.families[this.family.project]) {
-        this.family = null;
-      }
-
-      this.loadingFamilies = false;
-    },
-
-    async getImagesInProject(val, showDeprecated) {
-      let imagesInProject = [];
-
-      try {
-        if (val?.family) {
-          imagesInProject = await getGKEImageFamilies(this.$store, this.credentialId, this.project, location, val.family, val.project, showDeprecated);
-        }
-      } catch (e) {
-        this.errors.push(e.data);
-      }
-
-      return imagesInProject;
-    },
-
-    async getImages(val) {
-      this.loadingImages = true;
-      try {
-        this.machineImages = await this.getImagesInProject(val, this.showDeprecated);
-        // If we had to reload list of images, we need to reset selected image if it is no longer in the list
-        if ( !!this.machineImage && this.machineImages.filter((image) => image.name === this.machineImage).length === 0) {
-          this.machineImage = '';
-        }
-      } catch (e) {
-        this.errors.push(e.data);
-      }
-
-      this.loadingImages = false;
     },
 
     async getDiskTypes() {
@@ -579,27 +359,11 @@ export default {
       }
       this.loadingMachineTypes = false;
     },
-    setMachineImage(neu) {
-      this.value.machineImage = this.formatMachineImage(neu?.selfLink);
-    },
-    formatMachineImage(image) {
-      if (!image) {
-        return '';
-      }
-      const index = image?.indexOf('/projects/');
-
-      if (index === -1 ) {
-        return '';
-      }
-
-      return image?.substring(index + '/projects/'.length);
-    },
 
     async getOptions() {
       await this.getDiskTypes();
       await this.getMachineTypes();
       await this.getNetworks();
-      await this.getFamilies();
       // These can finish loading later
       this.getSubnetworks();
       this.getSharedSubnetworks();
@@ -627,6 +391,7 @@ export default {
           color="error"
           :label="stringify(err)"
           :closable="true"
+          :data-testid="`gce-error-banner-${idx}`"
           @close="closeError(idx)"
         />
       </div>
@@ -640,61 +405,21 @@ export default {
           :options="zones"
 
           :loading="loadingZones"
-          data-testid="gke-zone-select"
+          data-testid="gce-zone-select"
           class="span-3 mr-10"
           required
         />
       </div>
-      <div class="row mt-20">
-        <div class="col span-6 mr-10">
-          <LabeledInput
-            v-model:value="imageProjects"
-            :mode="mode"
-            label-key="cluster.machineConfig.gce.imageProjects.label"
-            :placeholder="defaultProjects"
-            data-testid="gce-disk-size"
-            required
-            :rules="fvGetAndReportPathRules('imageProjects')"
-          />
-        </div>
-        <div>
-          <Checkbox
-            v-model:value="showDeprecated"
-            :mode="mode"
-            :label="t('cluster.machineConfig.gce.showDeprecated.label')"
-            class="mt-20"
-          />
-        </div>
-      </div>
-      <div class="row mt-20">
-        <LabeledSelect
-          v-model:value="family"
-          label-key="cluster.machineConfig.gce.family.label"
-          :mode="mode"
-          :options="familyOptions"
-          option-key="value"
-          option-label="label"
-          :loading="loadingZones"
-          data-testid="gke-zone-select"
-          class="span-3 mr-10"
-          :rules="fvGetAndReportPathRules('family')"
-        />
+      <GCEImage
+        v-model:value="value.machineImage"
+        :credentialId="credentialId"
+        :projectId="value.project"
+        :mode="mode"
+        :location="location"
+        @error="(val)=>errors.push(val)"
+        @validationChanged="(val)=>imageAreaValid=val"
+      />
 
-        <LabeledSelect
-          v-model:value="machineImage"
-          label-key="cluster.machineConfig.gce.machineImage.label"
-          :mode="mode"
-          :options="imageOptions"
-          option-key="value"
-          option-label="label"
-          :loading="loadingImages"
-          data-testid="gce-disk-type"
-          class="span-3"
-          :tooltip="t('cluster.machineConfig.gce.machineImage.tooltip')"
-          required
-          :rules="fvGetAndReportPathRules('machineImage')"
-        />
-      </div>
       <div class="row mt-20">
         <LabeledSelect
           v-model:value="value.diskType"
@@ -702,7 +427,7 @@ export default {
           :mode="mode"
           :options="diskTypes"
           :loading="loadingDiskTypes"
-          data-testid="gce-disk-type"
+          data-testid="gce-disk-type-select"
           required
           class="span-3 mr-10"
           :rules="fvGetAndReportPathRules('diskType')"
@@ -712,7 +437,7 @@ export default {
           :mode="mode"
           label-key="cluster.machineConfig.gce.diskSize.label"
           :placeholder="50"
-          data-testid="gce-disk-size"
+          data-testid="gce-disk-size-input"
           class="span-3 mr-10"
           required
           :rules="fvGetAndReportPathRules('diskSize')"
@@ -723,7 +448,7 @@ export default {
           :mode="mode"
           :options="machineTypes"
           :loading="loadingMachineTypes"
-          data-testid="gke-zone-select"
+          data-testid="gce-machine-type-select"
           required
           :rules="fvGetAndReportPathRules('machineType')"
         />
@@ -738,7 +463,7 @@ export default {
           option-key="name"
           option-label="label"
           :loading="loadingNetworks"
-          data-testid="gke-zone-select"
+          data-testid="gce-network-select"
           class="span-3 mr-10"
           required
           :rules="fvGetAndReportPathRules('network')"
@@ -752,7 +477,7 @@ export default {
           option-key="name"
           option-label="name"
           :loading="loadingNetworks"
-          data-testid="gke-zone-select"
+          data-testid="gce-subnetwork-select"
         />
       </div>
     </div>
@@ -764,7 +489,7 @@ export default {
           label-key="cluster.machineConfig.gce.username.label"
           :placeholder="t('cluster.machineConfig.gce.username.placeholder')"
           :tooltip="t('cluster.machineConfig.gce.username.tooltip')"
-          data-testid="gce-disk-size"
+          data-testid="gce-username-input"
           class="span-3 mr-10"
         />
 
@@ -774,7 +499,7 @@ export default {
           label-key="cluster.machineConfig.gce.address.label"
           :placeholder="t('cluster.machineConfig.gce.address.placeholder')"
           :tooltip="t('cluster.machineConfig.gce.address.tooltip')"
-          data-testid="gce-disk-size"
+          data-testid="gce-address-input"
           class="span-3"
         />
       </div>
@@ -787,6 +512,7 @@ export default {
         :add-label="t('cluster.machineConfig.gce.scopes.add')"
         :disabled="disabled"
         class="col mt-20 span-10"
+        data-testid="gce-scopes-array"
       />
       <h3 class="mt-20">
         {{ t('cluster.machineConfig.gce.firewall.header') }}
@@ -798,10 +524,12 @@ export default {
             :mode="mode"
             :label="t('cluster.machineConfig.gce.internalFirewall.label')"
             :tooltip="t('cluster.machineConfig.gce.internalFirewall.tooltip')"
+            data-testid="gce-internal-firewall-prefix-checkbox"
           />
           <Banner
             color="info"
             label-key="cluster.machineConfig.gce.internalFirewall.banner"
+            data-testid="gce-internal-firewall-banner"
           />
           <ArrayList
             :value="tags"
@@ -809,6 +537,7 @@ export default {
             :title="t('gke.tags.label')"
             :add-label="t('gke.tags.add')"
             class="col span-6 mr-10"
+            data-testid="gce-tags-array"
           />
         </div>
         <div class="col span-6">
@@ -817,11 +546,13 @@ export default {
             :mode="mode"
             :label="t('cluster.machineConfig.gce.externalFirewall.label')"
             :tooltip="t('cluster.machineConfig.gce.externalFirewall.tooltip')"
+            data-testid="gce-external-firewall-prefix-checkbox"
           />
           <div v-if="!!value.setExternalFirewallRulePrefix">
             <Banner
               color="info"
               label-key="cluster.machineConfig.gce.externalFirewall.banner"
+              data-testid="gce-external-firewall-banner"
             />
             <ArrayList
               v-model:value="value.openPort"
@@ -829,6 +560,7 @@ export default {
               :title="t('cluster.machineConfig.gce.openPort.label')"
               :add-label="t('cluster.machineConfig.gce.openPort.add')"
               class="col span-6"
+              data-testid="gce-ports-array"
             />
           </div>
         </div>
@@ -843,7 +575,7 @@ export default {
           :value-can-be-empty="true"
           :add-label="t('aks.nodePools.labels.add')"
           :read-allowed="false"
-          data-testid="aks-node-labels-input"
+          data-testid="gce-labels-kv"
         />
       </div>
     </portal>
